get_history/requirements.txt
requests
python-dateutil
numpy==1.26.4
pandas==2.2.2
pyarrow==16.1.0
python-dotenv